Removed static references to SPI instances (backport from jsr354-api)
In a dynamic environment such as OSGI, SPI instances are registered
dynamically. The classes Monetary, MonetaryConversions and
MonetaryFormats keep static references to SPI instances. This breaks
usage of money-api in OSGI as SPI instances may be registered delayed.

To ensure that using money-api is in an OSGI environment is possible,
references to SPI instances should be resolved dynamically.

Added tests for Monetary, MonetaryFormats and MonetaryConversions.